pyproject.toml
[project]
name = "ak"
version = "0.1.0"
description = "Meeting management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
    "psycopg>=3.2.9",
    "psycopg2>=2.9.10",
    "python-decouple>=3.8",
    "requests>=2.32.4",
    "whitenoise>=6.5.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.2.0",
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
